update from vjoy:master (#4)
* GenInterface - Modifying Common API

Function SetDevPov(): Modified and tested
Function isDevOwned(): Modified and tested
Function isDevExist(): Modified and tested (+6 squashed commit)
Function GetDevId(): Modified and tested
Function GetDevNumber(): Modified and tested
Returns error when access to devices that were removed by another application
Function RelinquishDev(): Modified and tested
Function GetDevType(): Modified and tested (Known problem: Wrong results for devices that were forciblly removed)
ForceFree: Force relinquishment of vXbox devices
Function AcquireDev(): Modified and tested (+1 squashed commits)
GetDevType - Enhancement
Returns error when access to devices that were removed by another application (+1 squashed commits)
Function RelinquishDev(): Modified and tested
Function GetDevType(): Modified and tested (Known problem: Wrong results for devices that were forciblly removed) (+2 squashed commit)
 vGetInterface: Added a testing utility
ForceFree: Force relinquishment of vXbox devices
Function AcquireDev(): Modified and tested

* GenInterface - Modifying vJoy Compatible API

Improved tester CompatTest
Function GetVJDStatus(): Modified and tested

* Administrative Work

* Fixed compilation errors

* Eliminate the need to include file public.h

* Documentation for vGenInterface

* Bug fix: DLLs were not compiled correctly

Added "Extern "C""
Changed vXbox API function from SetBtn to SetButton
Updated doc

* Rename interface functions

IsAxisExists() ⇒ IsAxisExist()
isVBusExists() ⇒ isVBusExist()

* SetDevPov()

Change Range to 0-359.99

* Add interface function vJoyEnabled()

Was somehow left out

* isDevExist(): Interface function modified

* Work of vJoy Reset functions

Initialize stick when loading API

* Temporary changes

* Remove irrelevant code (+1 squashed commits)

Squashed commits:

[2146bf6] Bug Fix

* Fix analysis warnings

* Reset functions fixed

1. Default values are taken from the registry (if data exists) on creation of device database in vJoyInterface
2. When calling ResetPovs() & ResetButtons - the data for POVs/Buttons is taken from  registry (if data exists) and ORs with the rest of the data.

Known Problem:
Reset/Initialize of discrete POVs from the registry is not functional (+2 squashed commit)

* vJoy new version - 2.1.7

* Bug Fix

Reported as Issue#8 (https://github.com/shauleiz/vJoy/issues/8)

* Remove warning

* Update SDK

* Can enumerate passed devices that are exclusive

* Stability of API

GetHandleByIndex() returns more accurate values

* Create index.ftml

* Create index.html

* Delete index.ftml

* Update index.html

* Create Latest.php

* Administrative

Prevent unneeded builds when nothing changed

* Get Position from device

Implemented in Kernel - not tested
Method: Adde handler to code GET_POSITIONS in file rawpdo.c

* Get Position from device

Implemented in Kernel -  tested
Added internal API function GetDevPosition(BYTE id, PJOYSTICK_POSITION_V2 pPosition)

* Get Position from device

Sync_Position() implemented and used in AcquireVJD()
All initialization of position removed from vJoyDeviceEntry()
Partial testing

* Moved version to 2.1.8

* Added new Interface Function GetOwnerPid()

- Returns PID of process owning a vJoy device
- Added debuging window to vJoyFeeder (to be later removed)

* Updated documentation

* vJoyList: Adding a new utility

This utility will display the devices and their capabilities
Enhancement: Improved icon
Enhancement: Change size of dialog box
Colour code for Existing/Missing/Owned
Displays all 16 devices (+2 squashed commit)

* vJoyList

Add version and details

* vJoyFeeder - Removed temporary debug code

* Added MFC file needed for vJoyList

mfc120u.dll (x86)

* Added MFC file needed for vJoyList

mfc120u.dll (x64)

* vJoyList

Added to the automated batch build application (BuildAll.bat)
Added to the installation

* Discrete POV Initialization

Bug fix - partially tested

* Improved logging

Solved issue #13

* GetOwnerPid() - Added to C# wrapper

Tested

* GetOwnerPid() - Added to C# wrapper

Documentation update

* Improving code quality

* Improving code quality

Solved C28160 problem in function ReAllocatePoolWithTag()

* Improving code quality

All 'status' returned by WdfRequestRetrieveOutputBuffer() are now examined

* Improving code quality

Overriding analyzer bug in analyzing function SetupDiGetDeviceRegistryProperty()

* Improving code quality

Driver.c: Replacing wcspcy() by safer wcspcy_s()

* Improving code quality

vJoyInterface.cpp: Replacing many parameters in _ftprintf_s to conform to the correct type (mainly %d==>%u)

* Improving code quality

Clean-up of installation script

* Bug fix: System halts when driver unloads

Problem is when calling WdfCollectionRemove(vJoyDeviceCollection, Device);
in function vJoyEvtDeviceContextCleanup()  (hid.c:line 1215)

* Improve FFB sync

The part in FfbWaitForData() that got the overlapped data was incorrect and could have created a deadlock

* Bug fixes

Memory Pool allocation
Graceful exit after failure to allocate memory pools.

* Bug Fix

FfbWaitForData(): Crashed because tried to close event handle that was not owned

* Upgraded environment

Moved to WDK 10.0.14393.0

* Installer

Welcome page: Re-introduced.
Copyright date: 2015=>2016
Minumum target: XP=>Vista

* Minor change in the order of  functions

Now Sync_Position() comes before FfbStartThread()

* Added file vGenInterface.dll (x86) to installation

* Added file vGenInterface.dll (x64) to installation

* SDK lib files refreshed

* Bug fix

Description: http://vjoystick.sourceforge.net/site/index.php/forum/4-Help/1228-fast-feeder-causes-spurious-inputs#4128
Solution: Set collection ID only when the rest of the collection is valid

* Correct completion code

vJoyGetPositionData() succeeds only if data copied to output buffer successfully
Completion code for handler for IOCTL_HID_READ_REPORT is now STATUS_UNSUCCESSFUL if function vJoyGetPositionData() fails.

* Improved efficiency

For each device: Added a per-device PositionReady (BOOL) to indicate that position data is unread.

Restrict reading IOCTL_HID_READ_REPORT requests to position data of devices that:
1. Exist (pDevContext->DeviceImplemented[id - 1])
2. Comtain fresh position data (pDevContext->PositionReady[id - 1])

* Bug Fix - FFB Spring Effect

The spring efect was missing from the registry because it was commented out

* Minor change in versioning tool

* Create LICENSE,txt

* Create LICENSE.txt

* Delete LICENSE,txt

* Update README.md

Typo

+++ apps/vGenInterface/vGenTest/vXboxTest/vXboxTest.cpp
@@ -1,8 +1,7 @@
 #define _CRT_SECURE_NO_WARNINGS
 
-#include "../../stdafx.h"
+#include "stdafx.h"
 #include <Xinput.h>
-#include "../../../../inc/public.h"
 #include "../../vGenInterface.h"
 
 void TopLoop(UINT iDev);
@@ -290,7 +289,7 @@ void PressDpad(UINT iDev)
 			continue;
 		}
 
-		for (int i = 0; i < len; i++)
+		for (size_t i = 0; i < len; i++)
 		{
 			if (tolower(Side[i]) == 'u')
 				Flags |= XINPUT_GAMEPAD_DPAD_UP;
